intro_fastmcp/simplecalc/pyproject.toml
[project]
name = "simplecalc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Manuel Gil", email = "manuelgilsitio@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
simplecalc = "simplecalc:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
